// bridge/typescript/src/_generated/dir/symbol/symbol.ts
export type SymbolKind = "variable" | "import" | "class" | "struct" | "interface" | "newtypeInterface" | "enum" | "enumField" | "function" | "label" | "extension" | "typeAlias" | "genericTypeParameter" | "genericValueParameter" | "associatedType" | "associatedConst" | "newtype";

export const SymbolKind = {
    /** Encode this value. */
    encode(writer: BinaryWriter, value: SymbolKind): void {
        encodeSymbolKind(writer, value);
    },

    /** Decode one SymbolKind. */
    decode(reader: BinaryReader): SymbolKind {
        return decodeSymbolKind(reader);
    },

    /** Return this value as JSON. */
    toJson(value: SymbolKind): Json {
        return toJsonSymbolKind(value);
    },

    /** Return one SymbolKind from one JSON value. */
    fromJson(value: Json): SymbolKind {
        return fromJsonSymbolKind(value);
    },
};
// ...
export type SymbolSpace = "type" | "value" | "label";

export const SymbolSpace = {
    /** Encode this value. */
    encode(writer: BinaryWriter, value: SymbolSpace): void {
        encodeSymbolSpace(writer, value);
    },

    /** Decode one SymbolSpace. */
    decode(reader: BinaryReader): SymbolSpace {
        return decodeSymbolSpace(reader);
    },

    /** Return this value as JSON. */
    toJson(value: SymbolSpace): Json {
        return toJsonSymbolSpace(value);
    },

    /** Return one SymbolSpace from one JSON value. */
    fromJson(value: Json): SymbolSpace {
        return fromJsonSymbolSpace(value);
    },
};